cultivation-logger/pyproject.toml
[project]
name = "cultivation-logger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python>=4.11.0.86",
    "python-decouple>=3.8",
    "requests>=2.32.3",
]
